Moves Dashboard to SampleApp
